move chunk tests
